+++ assets/javascripts/kitten/components/typography/title/stories.js
@@ -1,9 +1,10 @@
 import React from 'react'
-import { withKnobs, text, boolean, select } from '@storybook/addon-knobs'
+import { withKnobs, text, boolean, select, color } from '@storybook/addon-knobs'
 import { Title } from './index'
 import { Marger } from '../../layout/marger'
 import { Container } from '../../grid/container'
 import { Text } from '../../typography/text'
+import COLORS from '../../../constants/colors-config'
 
 const modifierOptions = {
   Primary: 'primary',
@@ -32,6 +33,7 @@ export const DefaultProps = () => (
         modifier={select('Modifier', modifierOptions, 'primary')}
         italic={boolean('Italic', false)}
         margin={boolean('Margin', false)}
+        cssColor={color('cssColor', COLORS.font1)}
       >
         {text('Title', 'Lorem ipsum dolor sit amet')}
       </Title>